+++ lib/value_struct/no_clone.rb
@@ -1,4 +1,4 @@
-module ValueStruct::Clone
+module ValueStruct::NoClone
   def clone
     self
   end