+++ ChangeLog.md
@@ -1,3 +1,8 @@
+### 0.7.0
+* Improve mixin system
+* More logical class/model/inheritance logic
+* Rename Clone mixin to NoClone
+
 ### 0.6.0
 * Fix specs
 * Fix #dup